+++ src/config.ts
@@ -22,12 +22,15 @@ import {
 } from 'fhir-works-on-aws-persistence-ddb';
 import JsonSchemaValidator from 'fhir-works-on-aws-routing/lib/router/validation/jsonSchemaValidator';
 import HapiFhirLambdaValidator from 'fhir-works-on-aws-routing/lib/router/validation/hapiFhirLambdaValidator';
+import SubscriptionValidator from 'fhir-works-on-aws-routing/lib/router/validation/subscriptionValidator';
+import getAllowListedSubscriptionEndpoints from './subscriptions/allowList';
 import RBACRules from './RBACRules';
 import { loadImplementationGuides } from './implementationGuides/loadCompiledIGs';
 
-const { IS_OFFLINE, ENABLE_MULTI_TENANCY } = process.env;
+const { IS_OFFLINE, ENABLE_MULTI_TENANCY, ENABLE_SUBSCRIPTIONS } = process.env;
 
 const enableMultiTenancy = ENABLE_MULTI_TENANCY === 'true';
+const enableSubscriptions = ENABLE_SUBSCRIPTIONS === 'true';
 
 export const fhirVersion: FhirVersion = '4.0.1';
 const baseResources = fhirVersion === '4.0.1' ? BASE_R4_RESOURCES : BASE_STU3_RESOURCES;
@@ -65,6 +68,7 @@ const esSearch = new ElasticSearchService(
     undefined,
     { enableMultiTenancy },
 );
+
 const s3DataService = new S3DataService(dynamoDbDataService, fhirVersion, { enableMultiTenancy });
 
 const OAuthUrl =
@@ -72,66 +76,74 @@ const OAuthUrl =
         ? 'https://OAUTH2.com'
         : process.env.OAUTH2_DOMAIN_ENDPOINT;
 
-export const getFhirConfig = async (): Promise<FhirConfig> => ({
-    configVersion: 1.0,
-    productInfo: {
-        orgName: 'Organization Name',
-    },
-    auth: {
-        authorization: authService,
-        // Used in Capability Statement Generation only
-        strategy: {
-            service: 'OAuth',
-            oauthPolicy: {
-                authorizationEndpoint: `${OAuthUrl}/authorize`,
-                tokenEndpoint: `${OAuthUrl}/token`,
+export const getFhirConfig = async (): Promise<FhirConfig> => {
+    if (enableSubscriptions) {
+        const subscriptionAllowList = await getAllowListedSubscriptionEndpoints();
+        validators.push(
+            new SubscriptionValidator(esSearch, dynamoDbDataService, subscriptionAllowList, { enableMultiTenancy }),
+        );
+    }
+    return {
+        configVersion: 1.0,
+        productInfo: {
+            orgName: 'Organization Name',
+        },
+        auth: {
+            authorization: authService,
+            // Used in Capability Statement Generation only
+            strategy: {
+                service: 'OAuth',
+                oauthPolicy: {
+                    authorizationEndpoint: `${OAuthUrl}/authorize`,
+                    tokenEndpoint: `${OAuthUrl}/token`,
+                },
             },
         },
-    },
-    server: {
-        // When running serverless offline, env vars are expressed as '[object Object]'
-        // https://github.com/serverless/serverless/issues/7087
-        // As of May 14, 2020, this bug has not been fixed and merged in
-        // https://github.com/serverless/serverless/pull/7147
-        url:
-            process.env.API_URL === '[object Object]' || process.env.API_URL === undefined
-                ? 'https://API_URL.com'
-                : process.env.API_URL,
-        dynamicHostName: process.env.ENABLE_DYNAMIC_HOST_NAME === 'true',
-    },
-    validators,
-    profile: {
-        systemOperations: ['transaction'],
-        bundle: dynamoDbBundleService,
-        compiledImplementationGuides: loadImplementationGuides('fhir-works-on-aws-routing'),
-        systemHistory: stubs.history,
-        systemSearch: stubs.search,
-        bulkDataAccess: dynamoDbDataService,
-        fhirVersion,
-        genericResource: {
-            operations: ['create', 'read', 'update', 'delete', 'vread', 'search-type'],
-            fhirVersions: [fhirVersion],
-            persistence: dynamoDbDataService,
-            typeSearch: esSearch,
-            typeHistory: stubs.history,
+        server: {
+            // When running serverless offline, env vars are expressed as '[object Object]'
+            // https://github.com/serverless/serverless/issues/7087
+            // As of May 14, 2020, this bug has not been fixed and merged in
+            // https://github.com/serverless/serverless/pull/7147
+            url:
+                process.env.API_URL === '[object Object]' || process.env.API_URL === undefined
+                    ? 'https://API_URL.com'
+                    : process.env.API_URL,
+                dynamicHostName: process.env.ENABLE_DYNAMIC_HOST_NAME === 'true',
         },
-        resources: {
-            Binary: {
-                operations: ['create', 'read', 'update', 'delete', 'vread'],
+        validators,
+        profile: {
+            systemOperations: ['transaction'],
+            bundle: dynamoDbBundleService,
+            compiledImplementationGuides: loadImplementationGuides('fhir-works-on-aws-routing'),
+            systemHistory: stubs.history,
+            systemSearch: stubs.search,
+            bulkDataAccess: dynamoDbDataService,
+            fhirVersion,
+            genericResource: {
+                operations: ['create', 'read', 'update', 'delete', 'vread', 'search-type'],
                 fhirVersions: [fhirVersion],
-                persistence: s3DataService,
-                typeSearch: stubs.search,
+                persistence: dynamoDbDataService,
+                typeSearch: esSearch,
                 typeHistory: stubs.history,
             },
+            resources: {
+                Binary: {
+                    operations: ['create', 'read', 'update', 'delete', 'vread'],
+                    fhirVersions: [fhirVersion],
+                    persistence: s3DataService,
+                    typeSearch: stubs.search,
+                    typeHistory: stubs.history,
+                },
+            },
         },
-    },
-    multiTenancyConfig: enableMultiTenancy
-        ? {
-              enableMultiTenancy: true,
-              useTenantSpecificUrl: true,
-              tenantIdClaimPath: 'custom:tenantId',
-          }
-        : undefined,
-});
+        multiTenancyConfig: enableMultiTenancy
+            ? {
+                  enableMultiTenancy: true,
+                  useTenantSpecificUrl: true,
+                  tenantIdClaimPath: 'custom:tenantId',
+              }
+            : undefined,
+    };
+};
 
 export const genericResources = baseResources;